added fraction LED opacity gradient support

+++ include/effects/RepeatEffect.h
@@ -6,10 +6,10 @@ namespace ledpipelines::effects {
 class RepeatEffect : public WrapperEffect {
 private:
     int numRepeats = 1;
-    int repeatDistance = 0;
+    float repeatDistance = 0;
 
 public:
-    RepeatEffect(BaseLedPipelineStage *stage, int repeatDistance, int numRepeats = 0);
+    RepeatEffect(BaseLedPipelineStage *stage, float repeatDistance, int numRepeats = 0);
 
     void calculate(float startIndex, TemporaryLedData &tempData) override;
 };
